refactor FirebaseHelper into local repos
split FirebaseHelper into UserRepository, ItemRepository,
ConversationRepository and refactored dependent classes & tests
