+++ app/src/main/java/com/github/brugapp/brug/view_model/SignInViewModel.kt
@@ -3,6 +3,7 @@ package com.github.brugapp.brug.view_model
 import android.content.Intent
 import androidx.lifecycle.ViewModel
 import com.github.brugapp.brug.data.FirebaseHelper
+import com.github.brugapp.brug.data.UserRepository
 import com.github.brugapp.brug.di.sign_in.*
 import com.github.brugapp.brug.model.User
 import com.google.firebase.auth.AuthCredential
@@ -42,7 +43,7 @@ class SignInViewModel @Inject constructor(
     // return new Brug User from SignInAccount
     private fun createNewBrugUser(account: SignInAccount?): User? {
         if (account == null) return null
-        return FirebaseHelper.createUserInFirestoreIfAbsent(auth.uid, account)
+        return UserRepository.createUserInFirestoreIfAbsent(auth.uid, account)
 //        val firstName = account.firstName
 //        val lastName = account.lastName
 //        val email = account.email
